==> Implementation/4. Data Cleaning 2.ipynb <==
import pandas as pd
import numpy as np

import fuzzywuzzy
from fuzzywuzzy import process

from nltk.stem import PorterStemmer
ps = PorterStemmer()

training_data = pd.read_csv("Datasets\\Training_Data.csv", encoding= 'unicode_escape', keep_default_na=False)
testing_data = pd.read_csv("Datasets\\Testing_Data.csv", encoding= 'unicode_escape', keep_default_na=False)
warning_train_data = pd.read_csv("Datasets\\Warning_Train_Data.csv", encoding= 'unicode_escape', keep_default_na=False)
warning_test_data = pd.read_csv("Datasets\\Warning_Test_Data.csv", encoding= 'unicode_escape', keep_default_na=False)

unique_schools = ['abbottabad', 'abbotabad public shool', 'abdali school', 'aes school for girls', 'agha khan school', 'agha khan board',
                 'agha khan university', 'ahmed public school', 'air base inter college', 'aitchison college', 'ayesha bawany academy', 
                  'army public school','ajk board', 'aizar school', 'alhuda school', 'ali ali boys', 'allama iqbal public school', 'almurtaza school',
                  'american school', 'angels school', 'apsacs', 'army burnhall college for boys', 'army public school and college', 
                  'army public school', 'army public school and degree college for boys', 'army pubmic school for girls azam garrison', 'asf public school', 
                  'army public school mahfooz pura lahore cantt', 'attia higher secondary school', 'aziz bhatti college', 'babar english school', 'bahawalpur', 'bahria college',
                  'beacon house school system', 'bahria foundation', 'bahria college', 'bahria college karsaz habib rahmat ullah road karachi', 'balochistan', 'balochistan residential college loralai', 
                  'beaconhouse canal side', 'beaconhouse garden town ', 'bise', 'blessing home', 'bloom field hall',
                 'bahria college naval complex e islamabad', 'blue horizon', 'bright future school', 'british council', 'british international school', 'bvs parsi high school abdullah haroon road',
                 'caa model school', 'cadet college', 'cambridge', 'cambridge international examinations', 'cathedral high school', 'central model school', 'chand bagh school', 
                 'city cliff school', 'city school', 'civil aviation', 'comprehensive model higher secondary school', 'connoiseur grammar school', 'convent of jesus and mary', 
                 'country grammar school', 'crescent model higher secondary school', 'crescent public', 'chinar army public school', 'chiniot islamia',
                 'customs public school', 'da high school', 'darearqam model high school', 'darululoom', 'dav public high school', 'defence authority model',
                 'defence public school', 'defence public school for boys', 'dera ghazi khan', 'dha khalifa bin zayad high school', 'district public school', 'divisional public school and college',
                 'don bosco school', 'dps model town', 'educators', 'edwards college', 'english grammar school', 'esena foundation',
                 'ever shine college', 'excelsior', 'f g school', 'faisal public high school', 'faisalabad', 'falcon house grammaer school',
                 'farooqi high school', 'fatimiyah boys school', 'fauji  foundation', 'fazaia degree college', 'fazaia inter college', 'federal government school',
                 'federal board of intermediate and secondary education', 'fg boys public school', 'forward model school', 'fouji foundation model school',
                 'foundation public school', 'frontier children academy', 'frontier model school', 'garrision boys high school', 'garrison acedemy for girls', 'ghazali school', 'giki institute school', 'glamour children academy',
                 'government school', 'govt comprehensive school', 'govt girls school', 'govt high school', 'govt muslim high school',
                 'govt central model school', 'govt pilot secondary school', 'grand folks paradise', 'greenwich', 'guards public school', 'govt higher secondary school',
                 'gujranwala', 'gulistan shah abdul latif', 'gulshan iqbal high school', 'habib public school', 'hamdard  public school',
                 'happy home high school', 'happy home english secondary school', 'happy palace school', 'heaven house high school', 'himayat e islam public school', 'hira high school',
                 'honours kinder garten', 'ibne sina', 'ideal', 'institute of learning and motivation', 'imtiaz public high school', 'international school',
                 'iqra huffaz school', 'islamabad college', 'islamia school', 'jinnah', 'kamran public school', 'karachi grammer secondary school',
                 'karachi public school', 'karakuram international', 'khanewal public school', 'kips high school for boys', 'kohinoor grammar school',
                 'kulsoom bai valika airport', 'la salle higher', 'laboratory high school', 'ladybird grammar school', 'lahore cadet school', 'lahore public school',
                 'lahore model school', 'lasalle high school', 'latif school', 'laural bank public school', 'layyah public school', 'lda model high school',
                 'little folks school', 'light house public school', 'mama parsi', 'masoomeen foundation', 'mazhar muslim model',
                 'mehran school', 'meritorius school', 'metropolis academy', 'metropolitan academy', 'millat public school', 'mirpurkhas',
                 'misali zakariya school', 'moon public school', 'mudrasa tul binat', 'muhammadi girls', 'multan public school',
                 'municipal high school', 'municipal inter college', 'muslim model high school', 'nasra school', 'national model high school', 'national high school', 'native public school', 'naveed foundation high school',
                 'new cantt public boys high school', 'new islamai public high school', 'new day secondary school', 'nishat high school',
                 'north school', 'noukhez public high school', 'nusrat jahan academy', 'opf public school', 'oxford school', 'paec model college',
                 'paf degree college', 'paf inter college', 'paf intermediate college', 'paf shaheen college', 'pak turk international school', 
                 'pakistan internation school', 'pakistan islamia higher secondary school', 'pakistan steel cadet college',
                 'parco public school', 'peshawar model school', 'peshawar public school', 'pilot secondary school', 'police public school', 'prince aly boys',
                 'pakistan international school', 'pakistan school', 'public school and college', 'quaid e azam public school',
                 'qurban boys high school', 'qurtuba public school', 'rahman children', 'roots montessori and high school', 'roshan tara model',
                 's m public academy campus', 'sacred heart cathederal school', 'saint anthonys', 'saint john', 'saint mary', 'saint paul',
                 'saint stephen', 'saleem nawaz college masroor', 'sardar khan high school', 'school of international studies', 'shaheen public school',
                 'shahwilayat public school', 'shakir school', 'shamsi society model school', 'sharif education complex', 'siqarah high school',
                 'sls montessori school nazimuddin', 'sms aga khan school', 'st anthonys high school', 'st bonaventure high school', 'st dominics convent school',
                 'st gregorys high school', 'st joseph convent', 'st marys high school khyber road peshawar', 'st michaels convent', 'st patricks school', 'standard model school',
                 'stpauls high school', 'sultan mohammad shah aga khan', 'superior science college', 'tabinda model high school', 'tameer e nasle nau high school quetta',
                 'the city school', 'the educational world secondary', 'the educators', 'the fahim school', 'the lahore lyceum', 'the mama parsi girls secondary',
                 'the model public school', 'the punjab school', 'the scholars school', 'the trust school', 'toronto school of academic', 'university of cambridge',
                 'university wensam college', 'usman science school', 'wapda boys high school', 'wapda cadet college', 'wapda senior model school',
                 'white house gammar school', 'zaheer public school', 'warsak model school', 'workers welfare school', 'unique high school', 'the punjab school',
                 'the educators progressive childrens academy', 'super radiant way high school', 'stmarys high school', 'st pauls english high school',
                 'super radiant way high school', 'saint marys academy', 'muhammadi girls', 'government central model school lower mall lahore', 'misali zakaria higher secondry school muzaffar garh',
                 'private', 'qamarebani hashim school', 'quaidiazam divisional public school gujranwala', 'sargodha']

unique_colleges = ['abbottabad', 'adamjee science college', 'adamjee govt science college', 'aga khan higher secondary school', 
                   'aga khan university', 'air base inter college ', 'aisha bawany', 'army burn hall college', 'army public school',
                   'army public school and college', 'bahawalpur', 'bahria colleg karsaz habib rahmatullah road', 'bahria college mt khan road',
                   'balochistan', 'balochistan residential college loralai',
                   'bamm pechs govt college for women', 'beacon house', 'bise dera ghazi khan', 'bise dg khan', 'bise faisalabad', 
                   'bise gujranwala', 'bise hyderabad', 'bise islamabad', 'bise karachi', 'bise lahore', 'bise  malakand', 'bise mirpur',
                   'bise quetta', 'bise rawalpindi', 'bise sargodha', 'bise sukkur', 'bise bahawalpur','bloomfield hall', 'british council karachi', 'british council peshawar',
                   'cadet college hasanabdal', 'cadet college petaro', 'cambridge', 'chenab college', 'city school' , 'college of emerging technologies',
                   'commecs', 'commecs institite of business and emerging sciences', 'community model', 'crescent college', 'cresent model',
                   'dj sindh govt science college', 'da degree college', 'defence degree college', 'defence public school and college',
                   'dehli college', 'dera ismail khan', 'dha degree college', 'divisional public school  and college', 'dps model town',
                   'dps and sahiwal', 'edexcel', 'edwards college', 'edwards school', 'emerson college', 'excelsior college', 'fg degree college',
                   'faizan high school tando allahyar', 'farhan college', 'fatimiyah college', 'fauji foundation college', 'fazaia degre college', 
                   'fazaia intermediate college', 'fbise', 'fc college', 'federal board', 'federal govt degree college', 'forman christian college',
                   'foundation public school', 'gc univesity', 'garrison science degree college', 'garrison academy', 'garrison post graduate college', 
                   'gift college', 'giki', 'global system of integrated studies', 'gojra college', 'gov islamia college civil lines',
                   'government college for boys', 'governement college women', 'government college of science',
                   'government delhi college', 'government post graduate college', 'govt college bosan road multan', 'govt college jampur', 'govt college university lahore',
                   'govt degree boys college gulzar e hijri', 'govt degree college chowk azam', 'govt degree college dg khan', 'govt degree college gulshan e iqbal',
                   'govt degree college hasilpur', 'govt degree college larkana', 'govt degree college malir cantt', 'govt girls high school tandoallahyar', 
                   'govt goru nanak degree college nankana sahib', 'gov higher secondary school', 'govt jahanzeb college saidu sharif swat',
                   'govt kinnaird college', 'govt model college', 'govt murray college sialkot', 'govt muslim college', 'govt national college', 'govt shalimar college',
                   'govt superior science college', 'govt zamindar college', 'guards public college',
                   'gulberg college', 'gulshan collage', 'hamdard college', 'hira public school sukkur', 'icms college system for boys', 'indus science college rawalpindi',
                   'international school and college of pakistan kuwait', 'islamia college'
                   'iqra college of science', 'islamabad model college', 'jinnah college for women',
                   'jinnah university college', 'kaim khani memorial high secondary  school tando allahyar', 'karachi foundation school', 
                   'kashmir model college mirpur', 'keynesian institute of management and sciences', 'khanewal public', 'khatoon e pakistan',
                   'fareed govt college', 'khyber college of sciences', 'kips intermediate college', 'kohat', 'lahore', 'beaconhouse defence',
                   'lahore college for women', 'lahore grammar school', 'larkana', 'layyah public higher secondry school', 'leadership college',                   
                   'london', 'manarat international school', 'masoomeen college chiniot', 'mazhar muslim model ranipur', 'mehran model school panoakil',
                   'memon academy high school shahdadpur', 'meritorious science college', 'midrassa tul binat', 'mirpur', 'misali', 'multan',
                   'muslim public higher secondary school multan', 'national college', 'national grammar school',
                   'ncr cet college', 'new fazaia degree college peshawar', 'nicon college', 'nisar shaheed college risalpur', 'nishat college of science',
                   'nishat science college', 'nusrat jehan inter college', 'opf boys college', 'others', 'paf college',
                   'pak turk international school and college', 'pakistan education center', 'pakistan college muscat',
                   'pakistan education foundation college', 'pakistan international school', 'pakistan islamia higher secondary school', 'pakistan scout cadet college batrasi mansehra',
                   'pakistan shipowner college', 'pechs education foundation govt college', 'pechs foundation govt college', 'pechs govt college for women',
                   'peshawar model degree college', 'peshawar public school and college', 'pis saudi arabia', 'preston intermediate college', 'private',
                   'govt degree college sanghar', 'public school hyderabad', 'punjab school', 'punjab college',
                   'punjab college for women', 'punjab college of commerce', 'punjab college of information technology', 'punjab group of collges',
                   'punjab technical', 'quaid e azam college', 'queen marry college', 'qurtuba college hayatabad', 'rawalpindi', 'rise college of science',
                   'roshan tara model higher secondary school', 'sadiq public school', 'sm science college', 'sahiwal', 'saidu sharif',
                   'saint patricks', 'salamat icas', 'saudi arabian international', 'scholars college', 'shaheen academy higher secondary school',
                   'shaheen intermediate college', 'sindh', 'sindh madressah  tul islam', 'sir adamjee institute', 'sir syed govt girls college',
                   'sm govt science college', 'st josephs', 'st bonaventures', 'st lawrances girls collage', 'st patricks college', 'suffa educational complex',
                   'sukkur', 'summit education system', 'superior college faisalabad', 'superior college of science', 'superior science higher secondary school',
                   'tameer i nau college quetta', 'tcf higher sec school cowasjee', 'the avicenna school', 'the city school', 'the lyceum school',
                   'the mardan model school and college', 'superior college of commerce', 'unique college', 'univeristy of cambridge', 'university college lahore', 
                   'university wensam college', 'ustad bukhari degree college dadu', 'wapda cadet college tarbela', 'wapda degree college terbela', 'warsak model school and college',
                   'wensam college']

def replace_matches_in_column(df, column, string_to_match, min_ratio = 55):
    # get a list of unique strings
    strings = df[column].unique()
    
    # get the top 10 closest matches to our input string
    matches = fuzzywuzzy.process.extract(string_to_match, strings, 
                                         limit=10, scorer=fuzzywuzzy.fuzz.token_sort_ratio)

    # only get matches with a ratio > 90
    close_matches = [matches[0] for matches in matches if matches[1] >= min_ratio]

    # get the rows of all the close matches in our dataframe
    rows_with_matches = df[column].isin(close_matches)

    # replace all rows with close matches with the input matches 
    df.loc[rows_with_matches, column] = string_to_match

def preprocess(data, column, to_replace):
    data[column] = data[column].str.lower()
    data[column] = data[column].str.strip()
    data[column] = data[column].str.replace('[-./()#?*@\d,:]+', '')
    data.loc[data[column] == "", column] = 'others'
    data.loc[data[column] == "na", column] = 'others'
    data.loc[data[column] == "not given", column] = 'others'
    data.loc[data[column] == "not known", column] = 'others'
    data[column] = data[column].str.strip()
    
    preprocessed_list = []
    for i in data[column]:
        list_ = i.split()
        stemmed = []
        word = " "
        for j in list_:
            stemmed.append(ps.stem(j))
        preprocessed_list.append(word.join(stemmed))
        
    data[column] = preprocessed_list
    
    for i in to_replace:
        replace_matches_in_column(df=data, column=column, string_to_match=i)
        
    return data

print("Training data [School]", len(training_data["SCHOOL"].unique()))
print("Training data [College]", len(training_data["COLLEGE"].unique()))

print("Testing data [School]", len(testing_data["SCHOOL"].unique()))
print("Testing data [College]", len(testing_data["COLLEGE"].unique()))

print("Testing data [School]", len(warning_train_data["SCHOOL"].unique()))
print("Testing data [College]", len(warning_train_data["COLLEGE"].unique()))

print("Warning_Train_Data [School]", len(warning_test_data["SCHOOL"].unique()))
print("Warning_Test_Data [College]", len(warning_test_data["COLLEGE"].unique()))

training_data = preprocess(training_data, "SCHOOL", unique_schools)
training_data = preprocess(training_data, "COLLEGE", unique_colleges)

testing_data = preprocess(testing_data, "SCHOOL", unique_schools)
testing_data = preprocess(testing_data, "COLLEGE", unique_colleges)

warning_train_data = preprocess(warning_train_data, "SCHOOL", unique_schools)
warning_train_data = preprocess(warning_train_data, "COLLEGE", unique_colleges)

warning_test_data = preprocess(warning_test_data, "SCHOOL", unique_schools)
warning_test_data = preprocess(warning_test_data, "COLLEGE", unique_colleges)

print("Training data [School]", len(training_data["SCHOOL"].unique()))
print("Training data [College]", len(training_data["COLLEGE"].unique()))

print("Testing data [School]", len(testing_data["SCHOOL"].unique()))
print("Testing data [College]", len(testing_data["COLLEGE"].unique()))

print("Testing data [School]", len(warning_train_data["SCHOOL"].unique()))
print("Testing data [College]", len(warning_train_data["COLLEGE"].unique()))

print("Warning_Train_Data [School]", len(warning_test_data["SCHOOL"].unique()))
print("Warning_Test_Data [College]", len(warning_test_data["COLLEGE"].unique()))

training_data.to_csv("Final Data//Training_Data.csv", index=False)
testing_data.to_csv("Final Data//Testing_Data.csv", index=False)

warning_train_data.to_csv("Final Data//Warning_Train_Data.csv", index=False)
warning_test_data.to_csv("Final Data//Warning_Test_Data.csv", index=False)
