==> Implementation/5. Feature Selection 1.Rmd <==
---
title: "Feature Selection"
author: "Areeka Aijaz"
date: "5/10/2021"
output: html_document
---
  
```{r including external files, echo=FALSE, message=FALSE}
library(plyr)
library(dplyr)
library(ggplot2)

setwd("E:\\FYP\\FYP-1 (Part 2)\\Implementation")

```

```{r, echo=FALSE}

data <- read.csv("..\\..\\Final Data\\Training_Data.csv")

summary(data)

data$SCHOOL <- as.character(data$SCHOOL)
data$COLLEGE <- as.character(data$COLLEGE)

```

## Campus vs CGPA

Since there is imbalanced in no. of students, therefore selecting equal number of students randomly from each campus

```{r, echo=FALSE}

campus_detail <- data %>% group_by(CAMPUS) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(campus_detail)

sample_data <- ddply(data,.(data$CAMPUS),function(x) x[sample(nrow(x), 600),])

```

Now, applying ANOVA test to analyze the effect of CAMPUS on CGPA

```{r}
model <- aov(CGPA ~ CAMPUS, data = sample_data)
summary(model)
```

## Degree Programs vs CGPA

Since there is imbalanced in no. of students, therefore selecting equal number of students randomly from each program

```{r, echo=FALSE}

program_detail <- data %>% group_by(PROG_CODE) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(program_detail)

sample_data <- data %>% group_by(PROG_CODE) %>% filter(n() > 200)

sample_data <- ddply(sample_data,.(sample_data$PROG_CODE),function(x) x[sample(nrow(x), 200),])

```

Now, applying ANOVA test to analyze the effect of Degree Program on CGPA

```{r}
model <- aov(CGPA ~ PROG_CODE, data = sample_data)
summary(model)
```

## Gender vs CGPA

```{r, echo=FALSE}

gender_detail <- data %>% group_by(GENDER) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(gender_detail)

sample_data <- ddply(data,.(data$GENDER),function(x) x[sample(nrow(x), 3000),])

```

Now, applying ANOVA test to analyze the effect of Gender on CGPA

```{r}
model <- aov(CGPA ~ GENDER, data = sample_data)
summary(model)
```

## City vs CGPA

```{r, echo=FALSE}

sample_data <- data %>% group_by(CITY) %>% filter(n() > 60)

city_detail <- sample_data %>% group_by(CITY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(city_detail)

sample_data <- ddply(sample_data,.(sample_data$CITY),function(x) x[sample(nrow(x), 60),])

```

Now, applying ANOVA test to analyze the effect of City on CGPA

```{r}
model <- aov(CGPA ~ CITY, data = sample_data)
summary(model)
```

## SECONDARY vs CGPA

```{r, echo=FALSE}

sample_data <- data %>% filter(SECONDARY != "OTHER")

sec_detail <- sample_data %>% group_by(SECONDARY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(sec_detail)

sample_data <- ddply(sample_data,.(sample_data$SECONDARY),function(x) x[sample(nrow(x), 2000),])

```

Now, applying ANOVA test to analyze the effect of Secondary on CGPA

```{r}
model <- aov(CGPA ~ SECONDARY, data = sample_data)
summary(model)
```

## SCHOOL vs CGPA

```{r, echo=FALSE}

sample_data <- data %>% group_by(SCHOOL) %>% filter(n() > 20)

school_detail <- sample_data %>% group_by(SCHOOL) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(school_detail)

sample_data <- ddply(sample_data,.(sample_data$SCHOOL),function(x) x[sample(nrow(x), 20),])

```

Now, applying ANOVA test to analyze the effect of School on CGPA

```{r}
model <- aov(CGPA ~ SCHOOL, data = sample_data)
summary(model)
```

## HIGHER SECONDARY vs CGPA

```{r, echo=FALSE}

sample_data <- data %>% filter(HIGHER_SECONDARY != "OTHER")

sec_detail <- sample_data %>% group_by(HIGHER_SECONDARY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(sec_detail)

sample_data <- ddply(sample_data,.(sample_data$HIGHER_SECONDARY),function(x) x[sample(nrow(x), 500),])

```

Now, applying ANOVA test to analyze the effect of Higher Secondary on CGPA

```{r}
model <- aov(CGPA ~ HIGHER_SECONDARY, data = sample_data)
summary(model)
```

## COLLEGE vs CGPA

```{r, echo=FALSE}

sample_data <- data %>% group_by(COLLEGE) %>% filter(n() > 20)

college_detail <- sample_data %>% group_by(COLLEGE) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(college_detail)

sample_data <- ddply(sample_data,.(sample_data$COLLEGE),function(x) x[sample(nrow(x), 20),])

```

Now, applying ANOVA test to analyze the effect of College on CGPA

```{r}
model <- aov(CGPA ~ COLLEGE, data = sample_data)
summary(model)
```

## FIRST_SEM vs CGPA

```{r, echo=FALSE}

sample_data <- data %>% group_by(FIRST_SEM) %>% filter(n() > 50)

fir_sem_detail <- sample_data %>% group_by(FIRST_SEM) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(fir_sem_detail)

sample_data <- ddply(sample_data,.(sample_data$FIRST_SEM),function(x) x[sample(nrow(x), 50),])

```

Now, applying ANOVA test to analyze the effect of First Semester on CGPA

```{r}
model <- aov(CGPA ~ FIRST_SEM, data = sample_data)
summary(model)
```

## Effect of School  Grades on CGPA

```{r, echo=FALSE}

data %>% ggplot(aes(SEC_GRADE, CGPA, color = SEC_GRADE)) + geom_point(size = 0.9, alpha = 0.5, show.legend = FALSE) + theme_bw() + scale_x_continuous(limits = c(200,850)) + scale_color_gradient(low = "#0091ff", high = "#f0650e") + stat_smooth(method = "lm", size = 1, show.legend = FALSE) + ggtitle("Effect of School Grades on CGPA") + xlab("SECONDARY GRADE")
```

```{r}
cor.test(data$SEC_GRADE, data$CGPA, method = "pearson")
cor.test(data$SEC_GRADE, data$CGPA, method = "kendall")
cor.test(data$SEC_GRADE, data$CGPA, method = "spearman")
```

## Effect of College Grades on CGPA

```{r, echo=FALSE}

data %>% ggplot(aes(HIG_SEC_GRADE, CGPA, color = HIG_SEC_GRADE)) + geom_point(size = 0.9, alpha = 0.5, show.legend = FALSE) + theme_bw() + scale_x_continuous(limits = c(200,1100)) + scale_color_gradient(low = "#0091ff", high = "#f0650e") + stat_smooth(method = "lm", size = 1, show.legend = FALSE) + ggtitle("Effect of College Grades on CGPA") + xlab("HIGHER SECONDARY GRADE")

```

```{r}
cor.test(data$HIG_SEC_GRADE, data$CGPA, method = "pearson")
cor.test(data$HIG_SEC_GRADE, data$CGPA, method = "kendall")
cor.test(data$HIG_SEC_GRADE, data$CGPA, method = "spearman")
```

## Last Semester vs CGPA

```{r, echo=FALSE}

sample_data <- data %>% group_by(LAST_SEM) %>% filter(n() > 50)
  
las_sem_detail <- sample_data %>% group_by(LAST_SEM) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(las_sem_detail)

sample_data <- ddply(sample_data,.(sample_data$LAST_SEM),function(x) x[sample(nrow(x), 50),])

```

Now, applying ANOVA test to analyze the effect of Last Semester on CGPA

```{r}
model <- aov(CGPA ~ LAST_SEM, data = sample_data)
summary(model)
```


## Campus-wise Feature Selection

### KARACHI Campus

```{r, echo=FALSE}
KARACHI_data <- data %>% filter(grepl("KHI", CAMPUS))
```

## Degree Programs vs CGPA

Since there is imbalanced in no. of students, therefore selecting equal number of students randomly from each program

```{r, echo=FALSE}

program_detail <- KARACHI_data %>% group_by(PROG_CODE) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(program_detail)

sample_data <- KARACHI_data %>% group_by(PROG_CODE) %>% filter(n() > 200)

sample_data <- ddply(sample_data,.(sample_data$PROG_CODE),function(x) x[sample(nrow(x), 200),])

```

Now, applying ANOVA test to analyze the effect of Degree Program on CGPA

```{r}
model <- aov(CGPA ~ PROG_CODE, data = sample_data)
summary(model)
```

## Gender vs CGPA

```{r, echo=FALSE}

gender_detail <- KARACHI_data %>% group_by(GENDER) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(gender_detail)

sample_data <- ddply(KARACHI_data,.(KARACHI_data$GENDER),function(x) x[sample(nrow(x), 600),])

```

Now, applying ANOVA test to analyze the effect of Gender on CGPA

```{r}
model <- aov(CGPA ~ GENDER, data = sample_data)
summary(model)
```

## City vs CGPA

```{r, echo=FALSE}

sample_data <- KARACHI_data %>% group_by(CITY) %>% filter(n() > 9)

city_detail <- sample_data %>% group_by(CITY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(city_detail)

```

Now, applying ANOVA test to analyze the effect of City on CGPA

```{r}
model <- aov(CGPA ~ CITY, data = KARACHI_data)
summary(model)
```

## SECONDARY vs CGPA

```{r, echo=FALSE}

sample_data <- KARACHI_data %>% filter(SECONDARY != 'OTHER')

sec_detail <- sample_data %>% group_by(SECONDARY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(sec_detail)

sample_data <- ddply(sample_data,.(sample_data$SECONDARY),function(x) x[sample(nrow(x), 500),])

```

Now, applying ANOVA test to analyze the effect of Secondary on CGPA

```{r}
model <- aov(CGPA ~ SECONDARY, data = sample_data)
summary(model)
```

## SCHOOL vs CGPA

```{r, echo=FALSE}

sample_data <- KARACHI_data %>% group_by(SCHOOL) %>% filter(n() > 5)

school_detail <- sample_data %>% group_by(SCHOOL) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(school_detail)

```

Now, applying ANOVA test to analyze the effect of School on CGPA

```{r}
model <- aov(CGPA ~ SCHOOL, data = sample_data)
summary(model)
```

## HIGHER SECONDARY vs CGPA

```{r, echo=FALSE}

sample_data <- KARACHI_data %>% filter(HIGHER_SECONDARY != "OTHER")

sec_detail <- sample_data %>% group_by(HIGHER_SECONDARY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(sec_detail)

sample_data <- ddply(sample_data,.(sample_data$HIGHER_SECONDARY),function(x) x[sample(nrow(x), 320),])

```

Now, applying ANOVA test to analyze the effect of Higher Secondary on CGPA

```{r}
model <- aov(CGPA ~ HIGHER_SECONDARY, data = sample_data)
summary(model)
```

## COLLEGE vs CGPA

```{r, echo=FALSE}

sample_data <- KARACHI_data %>% group_by(COLLEGE) %>% filter(n() > 10)

college_detail <- sample_data %>% group_by(COLLEGE) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(college_detail)

```

Now, applying ANOVA test to analyze the effect of College on CGPA

```{r}
model <- aov(CGPA ~ COLLEGE, data = sample_data)
summary(model)
```

## FIRST_SEM vs CGPA

```{r, echo=FALSE}

sample_data <- KARACHI_data %>% group_by(FIRST_SEM) %>% filter(n() > 99)

fir_sem_detail <- sample_data %>% group_by(FIRST_SEM) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(fir_sem_detail)

```

Now, applying ANOVA test to analyze the effect of First Semester on CGPA

```{r}
model <- aov(CGPA ~ FIRST_SEM, data = sample_data)
summary(model)
```

## Effect of School  Grades on CGPA

```{r, echo=FALSE}

KARACHI_data %>% ggplot(aes(SEC_GRADE, CGPA, color = SEC_GRADE)) + geom_point(size = 0.9, alpha = 0.5, show.legend = FALSE) + theme_bw() + scale_x_continuous(limits = c(200,850)) + scale_color_gradient(low = "#0091ff", high = "#f0650e") + stat_smooth(method = "lm", size = 1, show.legend = FALSE) + ggtitle("Effect of School Grades on CGPA") + xlab("SECONDARY GRADE")
```

```{r}
cor.test(KARACHI_data$SEC_GRADE, KARACHI_data$CGPA, method = "pearson")
cor.test(KARACHI_data$SEC_GRADE, KARACHI_data$CGPA, method = "kendall")
cor.test(KARACHI_data$SEC_GRADE, KARACHI_data$CGPA, method = "spearman")
```

## Effect of College Grades on CGPA

```{r, echo=FALSE}

KARACHI_data %>% ggplot(aes(HIG_SEC_GRADE, CGPA, color = HIG_SEC_GRADE)) + geom_point(size = 0.9, alpha = 0.5, show.legend = FALSE) + theme_bw() + scale_x_continuous(limits = c(200,1100)) + scale_color_gradient(low = "#0091ff", high = "#f0650e") + stat_smooth(method = "lm", size = 1, show.legend = FALSE) + ggtitle("Effect of College Grades on CGPA") + xlab("HIGHER SECONDARY GRADE")

```

```{r}
cor.test(KARACHI_data$HIG_SEC_GRADE, KARACHI_data$CGPA, method = "pearson")
cor.test(KARACHI_data$HIG_SEC_GRADE, KARACHI_data$CGPA, method = "kendall")
cor.test(KARACHI_data$HIG_SEC_GRADE, KARACHI_data$CGPA, method = "spearman")
```

## Last Semester vs CGPA

```{r, echo=FALSE}

sample_data <- KARACHI_data %>% group_by(LAST_SEM) %>% filter(n() > 50)

las_sem_detail <- sample_data %>% group_by(LAST_SEM) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(las_sem_detail)

```

Now, applying ANOVA test to analyze the effect of Last Semester on CGPA

```{r}
model <- aov(CGPA ~ LAST_SEM, data = sample_data)
summary(model)
```

### ISLAMABAD Campus

```{r, echo=FALSE}
ISLAMABAD_data <- data %>% filter(grepl("ISB", CAMPUS))
```

## Degree Programs vs CGPA

Since there is imbalanced in no. of students, therefore selecting equal number of students randomly from each program

```{r, echo=FALSE}

program_detail <- ISLAMABAD_data %>% group_by(PROG_CODE) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(program_detail)

sample_data <- ISLAMABAD_data %>% group_by(PROG_CODE) %>% filter(n() > 200)

```

Now, applying ANOVA test to analyze the effect of Degree Program on CGPA

```{r}
model <- aov(CGPA ~ PROG_CODE, data = sample_data)
summary(model)
```

## Gender vs CGPA

```{r, echo=FALSE}

gender_detail <- sample_data %>% group_by(GENDER) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(gender_detail)

sample_data <- ddply(ISLAMABAD_data,.(ISLAMABAD_data$GENDER),function(x) x[sample(nrow(x), 1000),])

```

Now, applying ANOVA test to analyze the effect of Gender on CGPA

```{r}
model <- aov(CGPA ~ GENDER, data = sample_data)
summary(model)
```

## City vs CGPA

```{r, echo=FALSE}

sample_data <- ISLAMABAD_data %>% group_by(CITY) %>% filter(n() > 10)

city_detail <- sample_data %>% group_by(CITY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(city_detail)

```

Now, applying ANOVA test to analyze the effect of City on CGPA

```{r}
model <- aov(CGPA ~ CITY, data = sample_data)
summary(model)
```

## SECONDARY vs CGPA

```{r, echo=FALSE}


sample_data <- ISLAMABAD_data %>% filter(SECONDARY != "OTHER")

sec_detail <- sample_data %>% group_by(SECONDARY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(sec_detail)

sample_data <- ddply(sample_data,.(sample_data$SECONDARY),function(x) x[sample(nrow(x), 700),])

```

Now, applying ANOVA test to analyze the effect of Secondary on CGPA

```{r}
model <- aov(CGPA ~ SECONDARY, data = sample_data)
summary(model)
```

## SCHOOL vs CGPA

```{r, echo=FALSE}

sample_data <- ISLAMABAD_data %>% group_by(SCHOOL) %>% filter(n() > 5)

school_detail <- sample_data %>% group_by(SCHOOL) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(school_detail)

```

Now, applying ANOVA test to analyze the effect of School on CGPA

```{r}
model <- aov(CGPA ~ SCHOOL, data = sample_data)
summary(model)
```

## HIGHER SECONDARY vs CGPA

```{r, echo=FALSE}

sample_data <- ISLAMABAD_data %>% filter(HIGHER_SECONDARY != 'OTHER')

sec_detail <- sample_data %>% group_by(HIGHER_SECONDARY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(sec_detail)
  
sample_data <- ddply(sample_data,.(sample_data$HIGHER_SECONDARY),function(x) x[sample(nrow(x), 700),])

```

Now, applying ANOVA test to analyze the effect of Higher Secondary on CGPA

```{r}
model <- aov(CGPA ~ HIGHER_SECONDARY, data = sample_data)
summary(model)
```

## COLLEGE vs CGPA

```{r, echo=FALSE}

sample_data <- ISLAMABAD_data %>% group_by(COLLEGE) %>% filter(n() > 10)

college_detail <- sample_data %>% group_by(COLLEGE) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(college_detail)

```

Now, applying ANOVA test to analyze the effect of College on CGPA

```{r}
model <- aov(CGPA ~ COLLEGE, data = sample_data)
summary(model)
```

## FIRST_SEM vs CGPA

```{r, echo=FALSE}

sample_data <- ISLAMABAD_data %>% group_by(FIRST_SEM) %>% filter(n() > 99)

fir_sem_detail <- sample_data %>% group_by(FIRST_SEM) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(fir_sem_detail)

```

Now, applying ANOVA test to analyze the effect of First Semester on CGPA

```{r}
model <- aov(CGPA ~ FIRST_SEM, data = sample_data)
summary(model)
```

## Effect of School  Grades on CGPA

```{r, echo=FALSE}

ISLAMABAD_data %>% ggplot(aes(SEC_GRADE, CGPA, color = SEC_GRADE)) + geom_point(size = 0.9, alpha = 0.5, show.legend = FALSE) + theme_bw() + scale_x_continuous(limits = c(200,850)) + scale_color_gradient(low = "#0091ff", high = "#f0650e") + stat_smooth(method = "lm", size = 1, show.legend = FALSE) + ggtitle("Effect of School Grades on CGPA") + xlab("SECONDARY GRADE")
```

```{r}
cor.test(ISLAMABAD_data$SEC_GRADE, ISLAMABAD_data$CGPA, method = "pearson")
cor.test(ISLAMABAD_data$SEC_GRADE, ISLAMABAD_data$CGPA, method = "kendall")
cor.test(ISLAMABAD_data$SEC_GRADE, ISLAMABAD_data$CGPA, method = "spearman")
```

## Effect of College Grades on CGPA

```{r, echo=FALSE}

ISLAMABAD_data %>% ggplot(aes(HIG_SEC_GRADE, CGPA, color = HIG_SEC_GRADE)) + geom_point(size = 0.9, alpha = 0.5, show.legend = FALSE) + theme_bw() + scale_x_continuous(limits = c(200,1100)) + scale_color_gradient(low = "#0091ff", high = "#f0650e") + stat_smooth(method = "lm", size = 1, show.legend = FALSE) + ggtitle("Effect of College Grades on CGPA") + xlab("HIGHER SECONDARY GRADE")

```

```{r}
cor.test(ISLAMABAD_data$HIG_SEC_GRADE, ISLAMABAD_data$CGPA, method = "pearson")
cor.test(ISLAMABAD_data$HIG_SEC_GRADE, ISLAMABAD_data$CGPA, method = "kendall")
cor.test(ISLAMABAD_data$HIG_SEC_GRADE, ISLAMABAD_data$CGPA, method = "spearman")
```

## Last Semester vs CGPA

```{r, echo=FALSE}

sample_data <- ISLAMABAD_data %>% group_by(LAST_SEM) %>% filter(n() > 50)

las_sem_detail <- sample_data %>% group_by(LAST_SEM) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(las_sem_detail)

sample_data <- ddply(sample_data,.(sample_data$LAST_SEM),function(x) x[sample(nrow(x), 50),])

```

Now, applying ANOVA test to analyze the effect of Last Semester on CGPA

```{r}
model <- aov(CGPA ~ LAST_SEM, data = sample_data)
summary(model)
```

### LAHORE Campus

```{r, echo=FALSE}
LAHORE_data <- data %>% filter(grepl("LHR", CAMPUS))
```

## Degree Programs vs CGPA

Since there is imbalanced in no. of students, therefore selecting equal number of students randomly from each program

```{r, echo=FALSE}

program_detail <- LAHORE_data %>% group_by(PROG_CODE) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(program_detail)

```

Now, applying ANOVA test to analyze the effect of Degree Program on CGPA

```{r}
model <- aov(CGPA ~ PROG_CODE, data = LAHORE_data)
summary(model)
```

## Gender vs CGPA

```{r, echo=FALSE}

gender_detail <- LAHORE_data %>% group_by(GENDER) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(gender_detail)

sample_data <- ddply(LAHORE_data,.(LAHORE_data$GENDER),function(x) x[sample(nrow(x), 1200),])

```

Now, applying ANOVA test to analyze the effect of Gender on CGPA

```{r}
model <- aov(CGPA ~ GENDER, data = sample_data)
summary(model)
```

## City vs CGPA

```{r, echo=FALSE}

sample_data <- LAHORE_data %>% group_by(CITY) %>% filter(n() > 30)

city_detail <- sample_data %>% group_by(CITY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(city_detail)

sample_data <- ddply(sample_data,.(sample_data$CITY),function(x) x[sample(nrow(x), 30),])

```

Now, applying ANOVA test to analyze the effect of City on CGPA

```{r}
model <- aov(CGPA ~ CITY, data = sample_data)
summary(model)
```

## SECONDARY vs CGPA

```{r, echo=FALSE}

sample_data <- LAHORE_data %>% filter(SECONDARY != 'OTHER')

sec_detail <- sample_data %>% group_by(SECONDARY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(sec_detail)

sample_data <- ddply(sample_data,.(sample_data$SECONDARY),function(x) x[sample(nrow(x), 800),])

```

Now, applying ANOVA test to analyze the effect of Secondary on CGPA

```{r}
model <- aov(CGPA ~ SECONDARY, data = sample_data)
summary(model)
```

## SCHOOL vs CGPA

```{r, echo=FALSE}

sample_data <- LAHORE_data %>% group_by(SCHOOL) %>% filter(n() > 5)

school_detail <- sample_data %>% group_by(SCHOOL) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(school_detail)

```

Now, applying ANOVA test to analyze the effect of School on CGPA

```{r}
model <- aov(CGPA ~ SCHOOL, data = sample_data)
summary(model)
```

## HIGHER SECONDARY vs CGPA

```{r, echo=FALSE}

sample_data <- LAHORE_data %>% filter(HIGHER_SECONDARY != 'OTHER')

sec_detail <- sample_data %>% group_by(HIGHER_SECONDARY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(sec_detail)
  
sample_data <- ddply(sample_data,.(sample_data$HIGHER_SECONDARY),function(x) x[sample(nrow(x), 700),])

```

Now, applying ANOVA test to analyze the effect of Higher Secondary on CGPA

```{r}
model <- aov(CGPA ~ HIGHER_SECONDARY, data = sample_data)
summary(model)
```

## COLLEGE vs CGPA

```{r, echo=FALSE}

sample_data <- LAHORE_data %>% group_by(COLLEGE) %>% filter(n() > 10)

college_detail <- sample_data %>% group_by(COLLEGE) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(college_detail)

```

Now, applying ANOVA test to analyze the effect of College on CGPA

```{r}
model <- aov(CGPA ~ COLLEGE, data = sample_data)
summary(model)
```

## FIRST_SEM vs CGPA

```{r, echo=FALSE}

sample_data <- LAHORE_data %>% group_by(FIRST_SEM) %>% filter(n() > 99)

fir_sem_detail <- sample_data %>% group_by(FIRST_SEM) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(fir_sem_detail)

sample_data <- ddply(sample_data,.(sample_data$FIRST_SEM),function(x) x[sample(nrow(x), 99),])

```

Now, applying ANOVA test to analyze the effect of First Semester on CGPA

```{r}
model <- aov(CGPA ~ FIRST_SEM, data = sample_data)
summary(model)
```

## Effect of School  Grades on CGPA

```{r, echo=FALSE}

LAHORE_data %>% ggplot(aes(SEC_GRADE, CGPA, color = SEC_GRADE)) + geom_point(size = 0.9, alpha = 0.5, show.legend = FALSE) + theme_bw() + scale_x_continuous(limits = c(200,850)) + scale_color_gradient(low = "#0091ff", high = "#f0650e") + stat_smooth(method = "lm", size = 1, show.legend = FALSE) + ggtitle("Effect of School Grades on CGPA") + xlab("SECONDARY GRADE")
```

```{r}
cor.test(LAHORE_data$SEC_GRADE, LAHORE_data$CGPA, method = "pearson")
cor.test(LAHORE_data$SEC_GRADE, LAHORE_data$CGPA, method = "kendall")
cor.test(LAHORE_data$SEC_GRADE, LAHORE_data$CGPA, method = "spearman")
```

## Effect of College Grades on CGPA

```{r, echo=FALSE}

LAHORE_data %>% ggplot(aes(HIG_SEC_GRADE, CGPA, color = HIG_SEC_GRADE)) + geom_point(size = 0.9, alpha = 0.5, show.legend = FALSE) + theme_bw() + scale_x_continuous(limits = c(200,1100)) + scale_color_gradient(low = "#0091ff", high = "#f0650e") + stat_smooth(method = "lm", size = 1, show.legend = FALSE) + ggtitle("Effect of College Grades on CGPA") + xlab("HIGHER SECONDARY GRADE")

```

```{r}
cor.test(LAHORE_data$HIG_SEC_GRADE, LAHORE_data$CGPA, method = "pearson")
cor.test(LAHORE_data$HIG_SEC_GRADE, LAHORE_data$CGPA, method = "kendall")
cor.test(LAHORE_data$HIG_SEC_GRADE, LAHORE_data$CGPA, method = "spearman")
```

## Last Semester vs CGPA

```{r, echo=FALSE}

sample_data <- LAHORE_data %>% group_by(LAST_SEM) %>% filter(n() > 30)

las_sem_detail <- sample_data %>% group_by(LAST_SEM) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(las_sem_detail)
  
sample_data <- ddply(sample_data,.(sample_data$LAST_SEM),function(x) x[sample(nrow(x), 30),])

```

Now, applying ANOVA test to analyze the effect of Last Semester on CGPA

```{r}
model <- aov(CGPA ~ LAST_SEM, data = sample_data)
summary(model)
```

### PESHAWAR Campus

```{r, echo=FALSE}
PESHAWAR_data <- data %>% filter(grepl("PWR", CAMPUS))
```

## Degree Programs vs CGPA

Since there is imbalanced in no. of students, therefore selecting equal number of students randomly from each program

```{r, echo=FALSE}

program_detail <- PESHAWAR_data %>% group_by(PROG_CODE) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(program_detail)

sample_data <- PESHAWAR_data %>% group_by(PROG_CODE) %>% filter(n() > 70)

sample_data <- ddply(sample_data,.(sample_data$PROG_CODE),function(x) x[sample(nrow(x), 70),])

```

Now, applying ANOVA test to analyze the effect of Degree Program on CGPA

```{r}
model <- aov(CGPA ~ PROG_CODE, data = sample_data)
summary(model)
```

## Gender vs CGPA

```{r, echo=FALSE}

gender_detail <- PESHAWAR_data %>% group_by(GENDER) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(gender_detail)

sample_data <- ddply(PESHAWAR_data,.(PESHAWAR_data$GENDER),function(x) x[sample(nrow(x), 100),])

```

Now, applying ANOVA test to analyze the effect of Gender on CGPA

```{r}
model <- aov(CGPA ~ GENDER, data = sample_data)
summary(model)
```

## City vs CGPA

```{r, echo=FALSE}

sample_data <- PESHAWAR_data %>% group_by(CITY) %>% filter(n() > 10)
  
city_detail <- sample_data %>% group_by(CITY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(city_detail)

```

Now, applying ANOVA test to analyze the effect of City on CGPA

```{r}
model <- aov(CGPA ~ CITY, data = sample_data)
summary(model)
```

## SECONDARY vs CGPA

Only 2% of students of OLEVEL

```{r, echo=FALSE}

sample_data <- PESHAWAR_data %>% filter(SECONDARY != 'OTHER')

sec_detail <- sample_data %>% group_by(SECONDARY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(sec_detail)

sample_data <- ddply(sample_data,.(sample_data$SECONDARY),function(x) x[sample(nrow(x), 35),])

```

Now, applying ANOVA test to analyze the effect of Secondary on CGPA

```{r}
model <- aov(CGPA ~ SECONDARY, data = sample_data)
summary(model)
```

## SCHOOL vs CGPA

```{r, echo=FALSE}

sample_data <- PESHAWAR_data %>% group_by(SCHOOL) %>% filter(n() > 10)

school_detail <- sample_data %>% group_by(SCHOOL) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(school_detail)
  
sample_data <- ddply(sample_data,.(sample_data$SCHOOL),function(x) x[sample(nrow(x), 10),])

```

Now, applying ANOVA test to analyze the effect of School on CGPA

```{r}
model <- aov(CGPA ~ SCHOOL, data = sample_data)
summary(model)
```

## HIGHER SECONDARY vs CGPA

```{r, echo=FALSE}

sample_data <- PESHAWAR_data %>% filter(HIGHER_SECONDARY != 'OTHER')
 
sec_detail <- sample_data %>% group_by(HIGHER_SECONDARY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(sec_detail)

sample_data <- ddply(sample_data,.(sample_data$HIGHER_SECONDARY),function(x) x[sample(nrow(x), 30),])

```

Now, applying ANOVA test to analyze the effect of Higher Secondary on CGPA

```{r}
model <- aov(CGPA ~ HIGHER_SECONDARY, data = sample_data)
summary(model)
```

## COLLEGE vs CGPA

```{r, echo=FALSE}

sample_data <- PESHAWAR_data %>% group_by(COLLEGE) %>% filter(n() > 10)
  
sample_data <- ddply(sample_data,.(sample_data$COLLEGE),function(x) x[sample(nrow(x), 10),])

college_detail <- sample_data %>% group_by(COLLEGE) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(college_detail)

```

Now, applying ANOVA test to analyze the effect of College on CGPA

```{r}
model <- aov(CGPA ~ COLLEGE, data = sample_data)
summary(model)
```

## FIRST_SEM vs CGPA

```{r, echo=FALSE}

sample_data <- PESHAWAR_data %>% group_by(FIRST_SEM) %>% filter(n() > 99)

fir_sem_detail <- sample_data %>% group_by(FIRST_SEM) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(fir_sem_detail)

sample_data <- ddply(sample_data,.(sample_data$FIRST_SEM),function(x) x[sample(nrow(x), 99),])

```

Now, applying ANOVA test to analyze the effect of First Semester on CGPA

```{r}
model <- aov(CGPA ~ FIRST_SEM, data = sample_data)
summary(model)
```

## Effect of School  Grades on CGPA

```{r, echo=FALSE}

PESHAWAR_data %>% ggplot(aes(SEC_GRADE, CGPA, color = SEC_GRADE)) + geom_point(size = 0.9, alpha = 0.5, show.legend = FALSE) + theme_bw() + scale_x_continuous(limits = c(200,850)) + scale_color_gradient(low = "#0091ff", high = "#f0650e") + stat_smooth(method = "lm", size = 1, show.legend = FALSE) + ggtitle("Effect of School Grades on CGPA") + xlab("SECONDARY GRADE")
```

```{r}
cor.test(PESHAWAR_data$SEC_GRADE, PESHAWAR_data$CGPA, method = "pearson")
cor.test(PESHAWAR_data$SEC_GRADE, PESHAWAR_data$CGPA, method = "kendall")
cor.test(PESHAWAR_data$SEC_GRADE, PESHAWAR_data$CGPA, method = "spearman")
```

## Effect of College Grades on CGPA

```{r, echo=FALSE}

PESHAWAR_data %>% ggplot(aes(HIG_SEC_GRADE, CGPA, color = HIG_SEC_GRADE)) + geom_point(size = 0.9, alpha = 0.5, show.legend = FALSE) + theme_bw() + scale_x_continuous(limits = c(200,1100)) + scale_color_gradient(low = "#0091ff", high = "#f0650e") + stat_smooth(method = "lm", size = 1, show.legend = FALSE) + ggtitle("Effect of College Grades on CGPA") + xlab("HIGHER SECONDARY GRADE")

```

```{r}
cor.test(PESHAWAR_data$HIG_SEC_GRADE, PESHAWAR_data$CGPA, method = "pearson")
cor.test(PESHAWAR_data$HIG_SEC_GRADE, PESHAWAR_data$CGPA, method = "kendall")
cor.test(PESHAWAR_data$HIG_SEC_GRADE, PESHAWAR_data$CGPA, method = "spearman")
```

## Last Semester vs CGPA

```{r, echo=FALSE}

sample_data <- PESHAWAR_data %>% group_by(LAST_SEM) %>% filter(n() > 30)

las_sem_detail <- sample_data %>% group_by(LAST_SEM) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(las_sem_detail)

sample_data <- ddply(sample_data,.(sample_data$LAST_SEM),function(x) x[sample(nrow(x), 30),])

```

Now, applying ANOVA test to analyze the effect of Last Semester on CGPA

```{r}
model <- aov(CGPA ~ LAST_SEM, data = sample_data)
summary(model)
```

### FAISALABAD Campus

```{r, echo=FALSE}
FAISALABAD_data <- data %>% filter(grepl("FSD", CAMPUS))
```

## Degree Programs vs CGPA

Since there is imbalanced in no. of students, therefore selecting equal number of students randomly from each program

```{r, echo=FALSE}

program_detail <- FAISALABAD_data %>% group_by(PROG_CODE) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(program_detail)

```

Now, applying ANOVA test to analyze the effect of Degree Program on CGPA

```{r}
model <- aov(CGPA ~ PROG_CODE, data = FAISALABAD_data)
summary(model)
```

## Gender vs CGPA

```{r, echo=FALSE}

gender_detail <- FAISALABAD_data %>% group_by(GENDER) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(gender_detail)

sample_data <- ddply(FAISALABAD_data,.(FAISALABAD_data$GENDER),function(x) x[sample(nrow(x), 100),])

```

Now, applying ANOVA test to analyze the effect of Gender on CGPA

```{r}
model <- aov(CGPA ~ GENDER, data = sample_data)
summary(model)
```

## City vs CGPA

```{r, echo=FALSE}

sample_data <- FAISALABAD_data %>% group_by(CITY) %>% filter(n() > 5)
  
city_detail <- sample_data %>% group_by(CITY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(city_detail)

```

Now, applying ANOVA test to analyze the effect of City on CGPA

```{r}
model <- aov(CGPA ~ CITY, data = sample_data)
summary(model)
```

## SECONDARY vs CGPA

```{r, echo=FALSE}

sec_detail <- FAISALABAD_data %>% group_by(SECONDARY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(sec_detail)

sample_data <- ddply(FAISALABAD_data,.(FAISALABAD_data$SECONDARY),function(x) x[sample(nrow(x), 64),])

```

Now, applying ANOVA test to analyze the effect of Secondary on CGPA

```{r}
model <- aov(CGPA ~ SECONDARY, data = sample_data)
summary(model)
```

## SCHOOL vs CGPA

```{r, echo=FALSE}

sample_data <- FAISALABAD_data %>% group_by(SCHOOL) %>% filter(n() > 10)

school_detail <- sample_data %>% group_by(SCHOOL) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(school_detail)

sample_data <- ddply(sample_data,.(sample_data$SCHOOL),function(x) x[sample(nrow(x), 10),])

```

Now, applying ANOVA test to analyze the effect of School on CGPA

```{r}
model <- aov(CGPA ~ SCHOOL, data = sample_data)
summary(model)
```

## HIGHER SECONDARY vs CGPA

```{r, echo=FALSE}

sample_data <- FAISALABAD_data %>% filter(HIGHER_SECONDARY != 'OTHER')

sec_detail <- sample_data %>% group_by(HIGHER_SECONDARY) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(sec_detail)

sample_data <- ddply(sample_data,.(sample_data$HIGHER_SECONDARY),function(x) x[sample(nrow(x), 40),])

```

Now, applying ANOVA test to analyze the effect of Higher Secondary on CGPA

```{r}
model <- aov(CGPA ~ HIGHER_SECONDARY, data = sample_data)
summary(model)
```

## COLLEGE vs CGPA

```{r, echo=FALSE}

sample_data <- FAISALABAD_data %>% group_by(COLLEGE) %>% filter(n() > 10)

college_detail <- sample_data %>% group_by(COLLEGE) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(college_detail)

sample_data <- ddply(sample_data,.(sample_data$COLLEGE),function(x) x[sample(nrow(x), 10),])

```

Now, applying ANOVA test to analyze the effect of College on CGPA

```{r}
model <- aov(CGPA ~ COLLEGE, data = sample_data)
summary(model)
```

## FIRST_SEM vs CGPA

```{r, echo=FALSE}

sample_data <- FAISALABAD_data %>% group_by(FIRST_SEM) %>% filter(n() > 99)

fir_sem_detail <- sample_data %>% group_by(FIRST_SEM) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(fir_sem_detail)

```

Now, applying ANOVA test to analyze the effect of First Semester on CGPA

```{r}
model <- aov(CGPA ~ FIRST_SEM, data = sample_data)
summary(model)
```

## Effect of School  Grades on CGPA

```{r, echo=FALSE}

FAISALABAD_data %>% ggplot(aes(SEC_GRADE, CGPA, color = SEC_GRADE)) + geom_point(size = 0.9, alpha = 0.5, show.legend = FALSE) + theme_bw() + scale_x_continuous(limits = c(200,850)) + scale_color_gradient(low = "#0091ff", high = "#f0650e") + stat_smooth(method = "lm", size = 1, show.legend = FALSE) + ggtitle("Effect of School Grades on CGPA") + xlab("SECONDARY GRADE")
```

```{r}
cor.test(FAISALABAD_data$SEC_GRADE, FAISALABAD_data$CGPA, method = "pearson")
cor.test(FAISALABAD_data$SEC_GRADE, FAISALABAD_data$CGPA, method = "kendall")
cor.test(FAISALABAD_data$SEC_GRADE, FAISALABAD_data$CGPA, method = "spearman")
```

## Effect of College Grades on CGPA

```{r, echo=FALSE}

FAISALABAD_data %>% ggplot(aes(HIG_SEC_GRADE, CGPA, color = HIG_SEC_GRADE)) + geom_point(size = 0.9, alpha = 0.5, show.legend = FALSE) + theme_bw() + scale_x_continuous(limits = c(200,1100)) + scale_color_gradient(low = "#0091ff", high = "#f0650e") + stat_smooth(method = "lm", size = 1, show.legend = FALSE) + ggtitle("Effect of College Grades on CGPA") + xlab("HIGHER SECONDARY GRADE")

```

```{r}
cor.test(FAISALABAD_data$HIG_SEC_GRADE, FAISALABAD_data$CGPA, method = "pearson")
cor.test(FAISALABAD_data$HIG_SEC_GRADE, FAISALABAD_data$CGPA, method = "kendall")
cor.test(FAISALABAD_data$HIG_SEC_GRADE, FAISALABAD_data$CGPA, method = "spearman")
```

## Last Semester vs CGPA

```{r, echo=FALSE}

sample_data <- FAISALABAD_data %>% group_by(LAST_SEM) %>% filter(n() > 30)

las_sem_detail <- sample_data %>% group_by(LAST_SEM) %>% summarise(Student_Count = n(), Min = round(min(CGPA), digits = 2), First_Quartile = round(summary(CGPA)[2], digits = 2),  Median = round(median(CGPA), digits = 2), Mean = round(mean(CGPA), digits = 2), Third_Quartile = round(summary(CGPA)[5], digits = 2), Max = round(max(CGPA), digits = 2))

knitr::kable(las_sem_detail)

```

Now, applying ANOVA test to analyze the effect of Last Semester on CGPA

```{r}
model <- aov(CGPA ~ LAST_SEM, data = sample_data)
summary(model)
```
