==> Implementation/6. Model Building (without warning data).ipynb <==
import pandas as pd
import numpy as np

from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import PolynomialFeatures

from sklearn.model_selection import train_test_split, cross_val_score, StratifiedShuffleSplit

from sklearn.metrics import mean_squared_error, r2_score, accuracy_score

from sklearn.linear_model import LinearRegression, LogisticRegression
from sklearn.naive_bayes import GaussianNB
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.svm import SVC 
from sklearn.ensemble import VotingClassifier

train_data = pd.read_csv("Final Data\\Training_Data.csv")
test_data = pd.read_csv("Final Data\\Testing_Data.csv")
train_data.head()

def data_preprocessing(data):

    ## Encoding data
    labelencoder = LabelEncoder()

    labelencoder.fit(data['GENDER'])
    data['GENDER'] = labelencoder.transform(data['GENDER'])
    gender_class = list(labelencoder.classes_)

    labelencoder.fit(data['BATCH'])
    data['BATCH'] = labelencoder.transform(data['BATCH'])
    batch_class = list(labelencoder.classes_)

    labelencoder.fit(data['CAMPUS'])
    data['CAMPUS'] = labelencoder.transform(data['CAMPUS'])
    campus_class = list(labelencoder.classes_)

    labelencoder.fit(data['PROG_CODE'])
    data['PROG_CODE'] = labelencoder.transform(data['PROG_CODE'])
    prog_code_class = list(labelencoder.classes_)

    labelencoder.fit(data['FIRST_SEM'])
    data['FIRST_SEM'] = labelencoder.transform(data['FIRST_SEM'])
    first_sem_class = list(labelencoder.classes_)

    labelencoder.fit(data['LAST_SEM'])
    data['LAST_SEM'] = labelencoder.transform(data['LAST_SEM'])
    last_sem_class = list(labelencoder.classes_)

    labelencoder.fit(data['CITY'])
    data['CITY'] = labelencoder.transform(data['CITY'])
    city_class = list(labelencoder.classes_)

    labelencoder.fit(data['SECONDARY'])
    data['SECONDARY'] = labelencoder.transform(data['SECONDARY'])
    sec_class = list(labelencoder.classes_)

    labelencoder.fit(data['HIGHER_SECONDARY'])
    data['HIGHER_SECONDARY'] = labelencoder.transform(data['HIGHER_SECONDARY'])
    hig_sec_class = list(labelencoder.classes_)

    labelencoder.fit(data['SCHOOL'])
    data['SCHOOL'] = labelencoder.transform(data['SCHOOL'])
    school_class = list(labelencoder.classes_)

    labelencoder.fit(data['COLLEGE'])
    data['COLLEGE'] = labelencoder.transform(data['COLLEGE'])
    college_class = list(labelencoder.classes_)
    
    ## Rearranging columns
    data = data[['SEM_1_CGPA', 'SEM_2_CGPA', 'SEM_3_CGPA', 'SEM_4_CGPA', 'SEM_5_CGPA', 'SEM_6_CGPA', 'SEM_7_CGPA', 'TOTAL_SEM', 
                 'GENDER', 'BATCH', 'CAMPUS', 'PROG_CODE', 'FIRST_SEM', 'LAST_SEM', 'CITY', 'SECONDARY', 'SCHOOL', 'SEC_GRADE', 
                 'HIGHER_SECONDARY', 'COLLEGE', 'HIG_SEC_GRADE', 'CGPA']]
    
    return data

train = pd.DataFrame()
train = data_preprocessing(train_data)

test = pd.DataFrame()
test = data_preprocessing(test_data)
test = test.fillna(0)

# Regression 

X = train.iloc[:, :-1]
y = train.iloc[:, -1]

X = (X - X.min()) / (X.max() - X.min())

X_train, X_validate, y_train, y_validate = train_test_split(X, y, test_size=0.2, random_state=1)

print('Training Data : ', len(X_train))
print("Validation Data : ", len(X_validate))

##  Linear Regression

"""
Linear regression applied on all features i.e: 'SEM_1_CGPA', 'SEM_2_CGPA', 'SEM_3_CGPA', 'SEM_4_CGPA', 'SEM_5_CGPA',
'SEM_6_CGPA', 'SEM_7_CGPA', 'TOTAL_SEM', 'GENDER', 'BATCH', 'CAMPUS', 'PROG_CODE', 'FIRST_SEM', 'LAST_SEM', 'CITY', 
'SECONDARY', 'SCHOOL', 'SEC_GRADE', 'HIGHER_SECONDARY', 'COLLEGE', 'HIG_SEC_GRADE', 'CGPA'
"""

model1 = LinearRegression()
model1.fit(X_train, y_train)
y_pred = model1.predict(X_validate)
model1_acc = r2_score(y_pred, y_validate)
rmse_1 = np.sqrt(mean_squared_error(y_validate, y_pred))
print("R^2:", model1_acc)
print("RMSE:", rmse_1)

## Polynomial Regression 

degree = 2
polynomial_features = PolynomialFeatures(degree = degree)

model2 = LinearRegression()

X_train_ = polynomial_features.fit_transform(X_train)
model2.fit(X_train_, y_train)

X_validate_ = polynomial_features.fit_transform(X_validate)
y_pred = model2.predict(X_validate_)

model2_acc = r2_score(y_validate, y_pred)
rmse_2 = np.sqrt(mean_squared_error(y_validate, y_pred))
print("R^2:", model2_acc)
print("RMSE:", rmse_2)

# Classification 

labelencoder = LabelEncoder()
y = labelencoder.fit_transform(pd.cut(train['CGPA'], 5, retbins=True)[0])
print(list(labelencoder.classes_))

X_train, X_validate, y_train, y_validate = train_test_split(X, y, test_size=0.3, random_state=1)

## Gaussian Naive Bayes

model3 = GaussianNB()
model3.fit(X_train, y_train)
y_pred = model3.predict(X_validate)
model3_acc = accuracy_score(y_validate, y_pred)
print("Accuracy:", model3_acc)

## Logistic Regression

model4 = LogisticRegression(solver='sag', multi_class='multinomial')
model4.fit(X_train, y_train)
model4_acc = model4.score(X_validate, y_validate)
print("Accuracy:", model4_acc)

## K-Nearest Neighbour

for i in range(1, 10):
    model5 = KNeighborsClassifier(n_neighbors=i) 
    model5.fit(X_train, y_train) 
    y_pred = model5.predict(X_validate) 
    model5_acc = accuracy_score(y_validate, y_pred)
    print(f"{i}-NN - Accuracy Score: {model5_acc}") 

## Decision Tree Classifier

model6 = DecisionTreeClassifier(criterion = 'entropy', ccp_alpha = 0.015)
model6.fit(X_train, y_train)
y_pred = model6.predict(X_validate)
model6_acc = accuracy_score(y_validate, y_pred)
print("Accuracy: ", model6_acc)

## Random Forest Classifier

model7 = RandomForestClassifier(n_estimators=100, bootstrap = True, max_features = 'sqrt', criterion = 'entropy')
model7.fit(X_train, y_train)
y_pred = model7.predict(X_validate)
model7_acc = accuracy_score(y_validate, y_pred)
print("Accuracy: ", model7_acc)

## Support Vector Machine

model8 = SVC(kernel="linear", C=1)
model8.fit(X_train, y_train)
y_pred = model8.predict(X_validate)
model8_acc = accuracy_score(y_validate, y_pred)
print("Accuracy: ", model8_acc)

## Ensemble Classifier

final = VotingClassifier(estimators=[('gnb', model3), 
                                     ('lgr', model4), 
                                     ('knn', model5), 
                                     ('dt', model6), 
                                     ('rf', model7),
                                     ('svm', model8)], 
                         voting='hard')

final.fit(X_train, y_train)
y_pred = final.predict(X_validate)
final_acc = accuracy_score(y_validate, y_pred)
print("Accuracy: ", final_acc)

# Results

print("=========Accuracy==========\n")

print("Linear Regression")
print("R-Squared:", round(model1_acc, 3))
print("RMSE: ", round(rmse_1, 3))

print("\nPolynmial Regression with degree 2")
print("R-Squared:", round(model2_acc, 3))
print("RMSE: ", round(rmse_2, 3))

print("\nGaussian Naive Bayes:", round(model3_acc, 3))
print("Logistic Regression:", round(model4_acc, 3))
print("K - Nearest Neighbor:", round(model5_acc, 3))
print("Decision Tree Classifier:", round(model6_acc, 3))
print("Random Forest Classifier:", round(model7_acc, 3))
print("Support Vector Machine:", round(model8_acc, 3))
print("Ensemble Classifier:", round(final_acc, 3))

# Classification using Hold-Out Approach Repeated 10 times

CV = StratifiedShuffleSplit(n_splits = 10, test_size = 0.3, random_state = 0)

## Gaussian Naive Bayes

model3 = GaussianNB()
model3_acc = cross_val_score(model3, X_train, y_train, cv=CV)
print("Accuracy:", model3_acc)

## Logistic Regression

model4 = LogisticRegression(solver='sag', multi_class='multinomial')
model4_acc = cross_val_score(model4, X_train, y_train, cv=CV)
print("Accuracy: ", model4_acc)

## K-Nearest Neighbour

model5 = KNeighborsClassifier(n_neighbors=3)
model5_acc = cross_val_score(model5, X_train, y_train, cv=CV)
print("Accuracy: ", model5_acc)

## Decision Tree Classifier

model6 = DecisionTreeClassifier(criterion = 'entropy', ccp_alpha = 0.015)
model6_acc = cross_val_score(model6, X_train, y_train, cv=CV, n_jobs = -1)
print("Accuracy: ", model6_acc)

## Random Forest Classifier

model7 = RandomForestClassifier(n_estimators=100, bootstrap = True, max_features = 'sqrt', criterion = 'entropy')
model7_acc = cross_val_score(model7, X_train, y_train, cv=CV, n_jobs = -1)
print("Accuracy: ", model7_acc)

## Support Vector Machine

model8 = SVC(kernel="linear", C=1)
model8_acc = cross_val_score(model6, X_train, y_train, cv=CV, n_jobs = -1)
print("Accuracy: ", model8_acc)

## Ensemble Classifier

final = VotingClassifier(estimators=[('gnb', model3), 
                                     ('lgr', model4), 
                                     ('knn', model5), 
                                     ('dt', model6), 
                                     ('rf', model7),
                                     ('svm', model8)], 
                         voting='hard')

final_acc = cross_val_score(final, X_train, y_train, cv=CV, n_jobs = -1)
print("Accuracy: ", final_acc)

# Results of Classification using Hold-out Repeated 10 times

print("==========================Accuracy=========================\n")

print("Gaussian Naive Bayes:", np.around(model3_acc, 3))
print(f"Mean Score: {np.round(np.mean(model3_acc), 3)}    Standard Deviation: {np.round(np.std(model3_acc), 3)}")

print("\nLogistic Regression:", np.around(model4_acc, 3))
print(f"Mean Score: {np.round(np.mean(model4_acc), 3)}    Standard Deviation: {np.round(np.std(model4_acc), 3)}")

print("\nK - Nearest Neighbor:", np.around(model5_acc, 3))
print(f"Mean Score: {np.round(np.mean(model5_acc), 3)}    Standard Deviation: {np.round(np.std(model5_acc), 3)}")

print("\nDecision Tree Classifier:", np.around(model6_acc, 3))
print(f"Mean Score: {np.round(np.mean(model6_acc), 3)}    Standard Deviation: {np.round(np.std(model6_acc), 3)}")

print("\nRandom Forest Classifier:", np.around(model7_acc, 3))
print(f"Mean Score: {np.round(np.mean(model7_acc), 3)}    Standard Deviation: {np.round(np.std(model7_acc), 3)}")

print("\nSupport Vector Machine:", np.around(model8_acc, 3))
print(f"Mean Score: {np.round(np.mean(model8_acc), 3)}    Standard Deviation: {np.round(np.std(model8_acc), 3)}")

print("\nEnsemble Classifier:", np.around(final_acc, 3))
print(f"Mean Score: {np.round(np.mean(final_acc), 3)}     Standard Deviation: {np.round(np.std(final_acc), 3)}")

# Principal Component Analysis

from sklearn.decomposition import PCA

pca = PCA()
pca.fit(X_train)
print(pca.explained_variance_ratio_)
print(pca.explained_variance_)

# Predictive Modeling

X = train.iloc[:, :-1]
y = train.iloc[:, -1]

X_test = test.iloc[:, :-1]

X_ = X[['SEM_1_CGPA', 'TOTAL_SEM', 'GENDER', 'BATCH', 'CAMPUS','PROG_CODE', 'FIRST_SEM', 'LAST_SEM', 'CITY', 'SECONDARY', 'SCHOOL',
        'SEC_GRADE', 'HIGHER_SECONDARY', 'COLLEGE', 'HIG_SEC_GRADE']]

test_sem_2 = X_test[X_test["TOTAL_SEM"] == 1]
test_sem_2 = test_sem_2[['SEM_1_CGPA', 'TOTAL_SEM', 'GENDER', 'BATCH', 'CAMPUS','PROG_CODE', 'FIRST_SEM', 'LAST_SEM', 'CITY', 
                         'SECONDARY', 'SCHOOL','SEC_GRADE', 'HIGHER_SECONDARY', 'COLLEGE', 'HIG_SEC_GRADE']]

print("Predict CGPA of 2nd Semester\nTotal Records:", len(test_sem_2))

model = LinearRegression()
model.fit(X_, y)
y_pred = np.round(model.predict(test_sem_2), 2)
test_sem_2['SEM_2_CGPA'] = y_pred
test_sem_2.head()

X_ = X[['SEM_1_CGPA', 'SEM_2_CGPA', 'TOTAL_SEM', 'GENDER', 'BATCH', 'CAMPUS','PROG_CODE', 'FIRST_SEM', 'LAST_SEM', 'CITY', 
        'SECONDARY', 'SCHOOL', 'SEC_GRADE', 'HIGHER_SECONDARY', 'COLLEGE', 'HIG_SEC_GRADE']]

test_sem_3 = X_test[X_test["TOTAL_SEM"] == 2]
test_sem_3 = test_sem_3[['SEM_1_CGPA', 'SEM_2_CGPA', 'TOTAL_SEM', 'GENDER', 'BATCH', 'CAMPUS','PROG_CODE', 'FIRST_SEM', 
                         'LAST_SEM', 'CITY', 'SECONDARY', 'SCHOOL','SEC_GRADE', 'HIGHER_SECONDARY', 'COLLEGE', 'HIG_SEC_GRADE']]

print("\nPredict CGPA of 3rd Semester\nTotal Records:", len(test_sem_3))

model = LinearRegression()
model.fit(X_, y)
y_pred = np.round(model.predict(test_sem_3), 2)
test_sem_3['SEM_3_CGPA'] = y_pred
test_sem_3.head()

X_ = X[['SEM_1_CGPA', 'SEM_2_CGPA', 'SEM_3_CGPA', 'TOTAL_SEM', 'GENDER', 'BATCH', 'CAMPUS','PROG_CODE', 'FIRST_SEM', 
        'LAST_SEM', 'CITY', 'SECONDARY', 'SCHOOL', 'SEC_GRADE', 'HIGHER_SECONDARY', 'COLLEGE', 'HIG_SEC_GRADE']]

test_sem_4 = X_test[X_test["TOTAL_SEM"] == 3]
test_sem_4 = test_sem_4[['SEM_1_CGPA', 'SEM_2_CGPA', 'SEM_3_CGPA', 'TOTAL_SEM', 'GENDER', 'BATCH', 'CAMPUS','PROG_CODE', 
                         'FIRST_SEM', 'LAST_SEM', 'CITY', 'SECONDARY', 'SCHOOL','SEC_GRADE', 'HIGHER_SECONDARY', 'COLLEGE', 
                         'HIG_SEC_GRADE']]

print("\nPredict CGPA of 4th Semester\nTotal Records:", len(test_sem_4))

model = LinearRegression()
model.fit(X_, y)
y_pred = np.round(model.predict(test_sem_4), 2)
test_sem_4['SEM_4_CGPA'] = y_pred
test_sem_4.head()

X_ = X[['SEM_1_CGPA', 'SEM_2_CGPA', 'SEM_3_CGPA', 'SEM_4_CGPA', 'TOTAL_SEM', 'GENDER', 'BATCH', 'CAMPUS','PROG_CODE', 
        'FIRST_SEM', 'LAST_SEM', 'CITY', 'SECONDARY', 'SCHOOL', 'SEC_GRADE', 'HIGHER_SECONDARY', 'COLLEGE', 'HIG_SEC_GRADE']]

test_sem_5 = X_test[X_test["TOTAL_SEM"] == 4]
test_sem_5 = test_sem_5[['SEM_1_CGPA', 'SEM_2_CGPA', 'SEM_3_CGPA', 'SEM_4_CGPA', 'TOTAL_SEM', 'GENDER', 'BATCH', 'CAMPUS',
                         'PROG_CODE', 'FIRST_SEM', 'LAST_SEM', 'CITY', 'SECONDARY', 'SCHOOL','SEC_GRADE', 'HIGHER_SECONDARY', 
                         'COLLEGE', 'HIG_SEC_GRADE']]

print("\nPredict CGPA of 5th Semester\nTotal Records:", len(test_sem_5))

model = LinearRegression()
model.fit(X_, y)
y_pred = np.round(model.predict(test_sem_5), 2)
test_sem_5['SEM_5_CGPA'] = y_pred
test_sem_5.head()

X_ = X[['SEM_1_CGPA', 'SEM_2_CGPA', 'SEM_3_CGPA', 'SEM_4_CGPA', 'SEM_5_CGPA', 'TOTAL_SEM', 'GENDER', 'BATCH', 'CAMPUS',
        'PROG_CODE', 'FIRST_SEM', 'LAST_SEM', 'CITY', 'SECONDARY', 'SCHOOL', 'SEC_GRADE', 'HIGHER_SECONDARY', 'COLLEGE', 
        'HIG_SEC_GRADE']]

test_sem_6 = X_test[X_test["TOTAL_SEM"] == 5]
test_sem_6 = test_sem_6[['SEM_1_CGPA', 'SEM_2_CGPA', 'SEM_3_CGPA', 'SEM_4_CGPA', 'SEM_5_CGPA', 'TOTAL_SEM', 'GENDER', 'BATCH', 
                         'CAMPUS','PROG_CODE', 'FIRST_SEM', 'LAST_SEM', 'CITY', 'SECONDARY', 'SCHOOL','SEC_GRADE', 
                         'HIGHER_SECONDARY', 'COLLEGE', 'HIG_SEC_GRADE']]

print("\nPredict CGPA of 6th Semester\nTotal Records:", len(test_sem_6))

model = LinearRegression() 
model.fit(X_, y)
y_pred = np.round(model.predict(test_sem_6), 2)
test_sem_6['SEM_6_CGPA'] = y_pred
test_sem_6.head()

X_ = X[['SEM_1_CGPA', 'SEM_2_CGPA', 'SEM_3_CGPA', 'SEM_4_CGPA', 'SEM_5_CGPA', 'SEM_6_CGPA', 'TOTAL_SEM', 'GENDER', 'BATCH', 
        'CAMPUS', 'PROG_CODE', 'FIRST_SEM', 'LAST_SEM', 'CITY', 'SECONDARY', 'SCHOOL', 'SEC_GRADE', 'HIGHER_SECONDARY', 
        'COLLEGE', 'HIG_SEC_GRADE']]

test_sem_7 = X_test[X_test["TOTAL_SEM"] == 6]
test_sem_7 = test_sem_7[['SEM_1_CGPA', 'SEM_2_CGPA', 'SEM_3_CGPA', 'SEM_4_CGPA', 'SEM_5_CGPA', 'SEM_6_CGPA', 'TOTAL_SEM', 
                         'GENDER', 'BATCH', 'CAMPUS','PROG_CODE', 'FIRST_SEM', 'LAST_SEM', 'CITY', 'SECONDARY', 'SCHOOL',
                         'SEC_GRADE', 'HIGHER_SECONDARY', 'COLLEGE', 'HIG_SEC_GRADE']]

print("\nPredict CGPA of 7th Semester\nTotal Records:", len(test_sem_7))

model = LinearRegression()
model.fit(X_, y)
y_pred = np.round(model.predict(test_sem_7), 2)
test_sem_7['SEM_7_CGPA'] = y_pred
test_sem_7.head()

X_ = X[['SEM_1_CGPA', 'SEM_2_CGPA', 'SEM_3_CGPA', 'SEM_4_CGPA', 'SEM_5_CGPA', 'SEM_6_CGPA', 'SEM_7_CGPA', 'TOTAL_SEM', 'GENDER', 
        'BATCH', 'CAMPUS', 'PROG_CODE', 'FIRST_SEM', 'LAST_SEM', 'CITY', 'SECONDARY', 'SCHOOL', 'SEC_GRADE', 'HIGHER_SECONDARY', 
        'COLLEGE', 'HIG_SEC_GRADE']]

test_sem_8 = X_test[X_test["TOTAL_SEM"] == 7]
test_sem_8 = test_sem_8[['SEM_1_CGPA', 'SEM_2_CGPA', 'SEM_3_CGPA', 'SEM_4_CGPA', 'SEM_5_CGPA', 'SEM_6_CGPA', 'SEM_7_CGPA', 
                         'TOTAL_SEM', 'GENDER', 'BATCH', 'CAMPUS','PROG_CODE', 'FIRST_SEM', 'LAST_SEM', 'CITY', 'SECONDARY', 
                         'SCHOOL','SEC_GRADE', 'HIGHER_SECONDARY', 'COLLEGE', 'HIG_SEC_GRADE']]

print("\nPredict CGPA of 8th Semester\nTotal Records:", len(test_sem_8))

model = LinearRegression()
model.fit(X_, y)
y_pred = np.round(model.predict(test_sem_8), 2)
test_sem_8['SEM_8_CGPA'] = y_pred
test_sem_8.head()
